created a couple new functions
loadDWfile is target of further work

one issue that is troubling: consistency in number of columns down
the rows of the text file

in test.txt, number of columns (as scanned in be textscan) changes from
15 to 14 at line 883 - this corresponds to line where attenuation/file
markers end.  

does this cause problems at assigning values to proper columns in the
data{} cell array?  keep an eye on this!


added:
  readDWfileheader.m
  readDWfileinfo.m
modified:
  countTextFileLines.m
  loadDWfile.m
unknown:
  Spreadsheet_BBN_1perSec.txt
  Spreadsheet_FRA_1perSec.txt